+++ src/helix/eventsub/get_eventsub_subscriptions.rs
@@ -48,6 +48,7 @@ pub struct EventSubSubscriptions {
         since = "0.5.0",
         note = "on 2021-05-11, this will no longer be returned. Use max_total_cost instead"
     )]
+    #[serde(default)]
     /// Subscription limit for client id that made the subscription creation request.
     pub limit: Option<usize>,
     /// Array containing subscriptions.